#58 clickable links in podviewer #354
DocBrowser
DocBrowser::POD
 improve resolver
 work around more of Pod::Perldoc
 better discovery of title
 fix bug that would resolve Handle to Coro/Handle.pm or
   IO/Handle.pm  (recursive switch to perldoc now OFF)
   
			
Wx::DocBrowser
	clean the search text
	shuffle some methods and method names

Padre::Locale
    add system_iso639 method


git-svn-id: http://svn.perlide.org/padre/trunk/Padre@5503 bc93166f-e056-4eb9-8c59-0a0840647e1b

+++ lib/Padre/DocBrowser/POD.pm
@@ -5,10 +5,9 @@ use strict;
 use warnings;
 use Config     ();
 use IO::Scalar ();
-use File::Spec ();
-use File::Spec::Functions;
+use Params::Util      qw( _INSTANCE );
 use Pod::Simple::XHTML ();
-use Pod::Abstract; ();
+use Pod::Abstract ();
 use Padre::DocBrowser::document ();
 use File::Temp                  ();
 
@@ -24,7 +23,7 @@ sub provider_for {
 
 # uri schema like http:// pod:// blah://
 sub accept_schemes {
-	'pod', 'perldoc',;
+	'perldoc'
 }
 
 sub viewer_for {
@@ -32,44 +31,73 @@ sub viewer_for {
 }
 
 sub resolve {
-	my $self  = shift;
-	my $ref   = shift;
-	my $hints = shift;
-	my ( $fh, $tempfile ) = File::Temp::tempfile();
+	my $self = shift;
+	my $ref  = shift;
+	my $hints= shift;
 
-	my @args = (
-		'-u', '-r',
-		"-d$tempfile",
-		( exists $hints->{lang} )
-		? ( '-L', ( $hints->{lang} ) )
-		: (),
-		$ref
+	my $query = $ref;
+	
+	if ( _INSTANCE($ref , 'URI')  ) {
+	    $query = $ref->opaque;
+	}
+	my ($docname,$section) = split_link($query);
+	
+	# Put Pod::Perldoc to work on $query
+	my ($fh,$tempfile) = File::Temp::tempfile();
+
+	my @args = ('-u' , 
+	    "-d$tempfile",
+	    (exists $hints->{lang} )
+	        ? ('-L' , ($hints->{lang}) )
+	        : (),
+	    (exists $hints->{perlfunc})
+	        ? '-f'
+	        : (),
+	    $query
 	);
+
 	my $pd = Padre::DocBrowser::pseudoPerldoc->new( args => \@args );
-	$pd->process();
+	my $perldoc_return = eval { $pd->process() };
+	return unless -s $tempfile;
+
 	my $pa = Pod::Abstract->load_file($tempfile);
 	close $fh;
-	unlink($tempfile);
-
+	unlink( $tempfile );
+	
 	my $doc = Padre::DocBrowser::document->new( body => $pa->pod );
 	$doc->mimetype('application/x-pod');
 	my $title_from = $hints->{title_from_section} || 'NAME';
-	my ($name) = $pa->select("/head1[\@heading =~ {$title_from}]");
-	if ($name) {
-		my $text = $name->text;
-		my ($module) = $text =~ /([^\s]+)/g;
-		$doc->title($module);
+	my $name;
+	if (
+	      ($name) = $pa->select("/head1[\@heading =~ {$title_from}]" )
+            or
+              ($name) = $pa->select("/head1") 
+       )
+	 {
+	        my $text= $name->text;
+	        my ($module) = $text =~ /([^\s]+)/g;
+	        $doc->title( $module );
 	}
-	unless ( $pa->select('/pod') || $pa->select('/head1') ) {
-		warn "$ref has no pod";
+	elsif ( ($name) = $pa->select("//item" ) )
+    {
+	        my $text = $name->pod;
+	        my ($item) = $text =~ /=item\s+([^\s]+)/g;
+	        $doc->title( $item );
+	}
+	        
+	unless ( $pa->select('/pod') || 
+                     $pa->select('//item') ||
+                     $pa->select('//head1' )
+    ) 
+    {
+	        warn "$ref has no pod in" . $pa->ptree;
+	        # Unresolvable ?
+	        return;
+	}
 
-		# Unresolvable ?
-		return;
-	}
 	return $doc;
-
-	# Perldoc failed - Unresolvable
-	return;
+	
+	
 }
 
 sub generate {
@@ -93,7 +121,16 @@ sub render {
 	my $response = Padre::DocBrowser::document->new;
 	$response->body( ${ $out->sref } );
 	$response->mimetype('text/xhtml');
+	$response->title( $doc->title );
 	return $response;
+}
+
+
+# Utility function , really wants to be inside a class like
+# URI::perldoc ??
+sub split_link {
+        my $query = shift;
+        my($doc,$section) = $query =~ m|([^/]+)/?+(.*+)|;
 }
 
 1;
@@ -118,6 +155,56 @@ sub find_good_formatter_class {
 	return;
 }
 
+# Even worse than monkey patching , copy paste from Pod::Perldoc w/ edits
+# to avoid untrappable calls to 'exit'
+sub process {
+    # if this ever returns, its retval will be used for exit(RETVAL)
+
+    my $self = shift;
+    # TODO: make it deal with being invoked as various different things
+    #  such as perlfaq".
+  
+    return $self->usage_brief  unless  @{ $self->{'args'} };
+    $self->pagers_guessing;
+    $self->options_reading;
+    $self->aside(sprintf "$0 => %s v%s\n", ref($self), $self->VERSION);
+    $self->drop_privs_maybe;
+    $self->options_processing;
+    
+    # Hm, we have @pages and @found, but we only really act on one
+    # file per call, with the exception of the opt_q hack, and with
+    # -l things
+
+    $self->aside("\n");
+
+    my @pages;
+    $self->{'pages'} = \@pages;
+    if(    $self->opt_f) { @pages = ("perlfunc")               }
+    elsif( $self->opt_q) { @pages = ("perlfaq1" .. "perlfaq9") }
+    else                 { @pages = @{$self->{'args'}}; }
+
+    return $self->usage_brief  unless  @pages;
+
+    $self->find_good_formatter_class();
+    $self->formatter_sanity_check();
+
+    $self->maybe_diddle_INC();
+      # for when we're apparently in a module or extension directory
+    
+    my @found = $self->grand_search_init(\@pages);
+    return unless @found;
+    
+    if ($self->opt_l) {
+        print join("\n", @found), "\n";
+        return;
+    }
+
+    $self->tweak_found_pathnames(\@found);
+    $self->assert_closing_stdout;
+    return $self->page_module_file(@found)  if  $self->opt_m;
+
+    return $self->render_and_page(\@found);
+}
 1;
 
 # Copyright 2008-2009 The Padre development team as listed in Padre.pm.
